Improve performance with canvas
Canvas can be very bad on performance when used the _wrong_/not optimized way as being noted in #35 it added 10 seconds of loading whereby this should be around 0.2-0.5 seconds and can be improved further but this focused on the way that the canvas was being used.

It will now re-use the canvas(I'm not sure if the heightmap is further in the game being executed again, but if so it will definitely be a good call). 
Disable image smoothing.
Only call the expensive getImageData once.
Do a little bit of maths to get the correct pixel value.